Refactor interrupt-based general utility functions
